requirements.txt
langchain
openai
python-dotenv
langchain-openai
pymysql
fastapi
streamlit
sqlalchemy